modules/panel-reuse-detection/requirements.txt
nb_utils
loguru
tqdm

numpy
opencv-python
pillow

pyyaml 
 
